# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: baja-telemetry-api
  version: 1.0.0-draft
  description: |
    Ingestao e consulta de telemetria do Baja SAE (UnBaja).

    Recebe frames CAN crus do ESP32, decodifica em grandezas fisicas usando o mapa de sinais
    em `contracts/can/unbaja.dbc`, e armazena em serie temporal.

    RASCUNHO da Fase 0 - o contrato se estabiliza quando a Fase 1 rodar de ponta a ponta.

    Ver docs/03-protocolo-ingestao.md e docs/08-contrato-de-erros.md.
servers:
  - url: http://localhost:8080
    description: Desenvolvimento local
...
security:
  - apiKey: []
paths:
...
  /api/v1/sessions/{sessionId}/summary:
    get:
      tags: [consulta]
      summary: Maximo, media e duracao por sinal
      operationId: sessionSummary
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Resumo por sinal
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SignalSummary' }
        '404': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Chave por dispositivo. Implementada na Fase 4
  parameters:
    SessionId:
      name: sessionId
      in: path
      required: true
      schema: { type: string, example: 2026-08-24-teste-suspensao }
  responses:
    Problem:
      description: Erro no formato Problem Details (RFC 7807). Ver docs/08
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Problem:
      type: object
      description: RFC 7807, mais a extensao `retryable`
      required: [type, title, status, retryable]
      properties:
        type:
          type: string
          format: uri
          description: Identificador ESTAVEL do tipo. O firmware decide por ele, nunca pelo texto
          example: https://baja.unb.br/errors/batch-too-large
        title:    { type: string, example: Lote acima do limite }
        status:   { type: integer, example: 413 }
        detail:   { type: string, example: "O lote tem 8000 frames; o maximo e 5000." }
        instance: { type: string, example: /api/v1/ingest }
        batchId:  { type: string, format: uuid }
        retryable:
          type: boolean
          description: |
            EXTENSAO NOSSA. true = o firmware deve guardar o buffer e retentar com o MESMO
            batchId. false = retentar produz o mesmo erro; descartar ou corrigir.
            Existe para que o firmware nao precise embutir a tabela de codigos HTTP. Ver docs/08.
          example: false
...
    SignalSummary:
      type: object
      required: [signal, canId, min, max, avg, count, invalidCount]
      properties:
        signal: { type: string, example: rpm }
        unit:
          type: string
          nullable: true
          description: Lida do DBC, nao do banco. Nula se o sinal saiu do mapa desde entao
          example: rpm
        canId: { type: integer, example: 256 }
        min:   { type: number, example: 0 }
        max:   { type: number, example: 4820.25 }
        avg:
          type: number
          description: Media PONDERADA pelo numero de pontos de cada janela, nao media de medias
          example: 2913.7
        count:
          type: integer
          format: int64
          description: Todos os pontos, INCLUINDO os marcados como invalidos
        invalidCount:
          type: integer
          format: int64
          description: Pontos fora da faixa do DBC. Gravados e marcados, nao descartados
          example: 0
...